+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Critical exceptions now will cause consumer to stop instead of retrying without a break
 - #412 - Fix dry-inflector dependency lock in gemspec
 - #414 - Backport to 1.2 the delayed retry upon failure
+- #437 - Raw message is no longer added to params after ParserError raised
 
 ## 1.2.8
 - #408 - Responder Topic Lookup Bug on Heroku